added router check phone
