+++ src/app.module.ts
@@ -1,11 +1,10 @@
 import { TypeOrmModule } from '@db/typeorm.config';
-import { Module } from '@nestjs/common';
-import { ConfigModule } from './config.module';
 import { AuthModule } from '@entities/auth/auth.module';
-import { UserModule } from '@entities/users/users.module';
 import { MailModule } from '@entities/mail/mail.module';
 import { QuestionsModule } from '@entities/questions/questions.module';
-import { PhoneModule } from './entities/phone/phone.module';
+import { UserModule } from '@entities/users/users.module';
+import { Module } from '@nestjs/common';
+import { ConfigModule } from './config.module';
 
 @Module({
   imports: [
@@ -15,7 +14,6 @@ import { PhoneModule } from './entities/phone/phone.module';
     UserModule,
     MailModule,
     QuestionsModule,
-    PhoneModule,
   ],
   controllers: [],
   providers: [],
