+++ src/entities/auth/auth.controller.ts
@@ -1,4 +1,16 @@
-import { Controller } from '@nestjs/common';
+import { Body, Controller, Post } from '@nestjs/common';
+import { ApiResponse, ApiTags } from '@nestjs/swagger';
+import { AuthService } from './auth.service';
+import { PhoneDto } from './dto/phone.dto';
 
-@Controller('auth')
-export class AuthController {}
+@ApiTags('Authentication')
+@Controller('api/auth')
+export class AuthController {
+  constructor(private readonly authService: AuthService) {}
+
+  @ApiResponse({ status: 200, type: String, description: 'OK' })
+  @Post('check-phone')
+  async checkPhone(@Body() body: PhoneDto) {
+    return this.authService.checkPhone(body.phone);
+  }
+}
