+++ src/entities/auth/auth.service.ts
@@ -1,4 +1,17 @@
-import { Injectable } from '@nestjs/common';
+import { ConflictException, Injectable } from '@nestjs/common';
 
 @Injectable()
-export class AuthService {}
+export class AuthService {
+  // !
+  //  Поправлю после создания модели юзера
+  private readonly userService: any;
+  constructor() {}
+
+  async checkPhone(phone: string) {
+    const user = await this.userService.getUser(phone);
+    if (user) {
+      throw new ConflictException('Phone number already exists');
+    }
+    return 'OK';
+  }
+}
