--- src/entities/phone/phone.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { PhoneController } from './phone.controller';
-import { PhoneService } from './phone.service';
-
-@Module({
-  controllers: [PhoneController],
-  providers: [PhoneService]
-})
-export class PhoneModule {}
